+++ tests/testthat/test-definetemplate.R
@@ -3,11 +3,11 @@ library(testthat)
 
 context("test-definetemplate")
 
-test_that("mrgtemplate is deprecated", {
+test_that("mrgtemplate is deprecated [YSP-TEST-0032]", {
   expect_warning(mrgtemplate())
 })
 
-test_that("definetemplate replaces mrgtemplate", {
+test_that("definetemplate replaces mrgtemplate [YSP-TEST-0033]", {
   loc <- normalizePath(definetemplate())
   expect_true(dir.exists(loc))
   files <- c("_output.yml", "header.tex")