+++ tests/testthat/test-select.R
@@ -4,7 +4,7 @@ library(testthat)
 
 context("test-select")
 
-test_that("select column subset", {
+test_that("select column subset [YSP-TEST-0084]", {
   spec <- ys_help$spec()
   # spec2 <- ys_select(spec)
   # expect_identical(spec,spec2)
@@ -14,7 +14,7 @@ test_that("select column subset", {
   expect_error(ys_select(spec, kyle))
 })
 
-test_that("select with no matching names returns zero-length yspec", {
+test_that("select with no matching names returns zero-length yspec [YSP-TEST-0085]", {
   spec <- ys_help$spec()
   spec <- ys_select(spec, character(0))
   expect_length(spec, 0)