+++ NEWS.md
@@ -1,3 +1,17 @@
+# yspec 0.5.3
+
+- Fix bug where character `type` was inferred whenever `values` were character,
+  overrriding what the user specified; now, only infer when user doesn't 
+  specify `type` (#129).
+
+- Add `ys_col_note()` to generate a string of column names and definitions
+  based on `short` or `label`; use this when column names appear in a table 
+  and require explanation in a a footnote (#119, #120).
+
+- Put stories in yaml format; add script to build validation docs from the
+  yaml file (#123, #124).
+
+
 # yspec 0.5.2
 
 - Adds `ys_extend()` to extend a spec object with additional